requirements.txt
flask==3.0.0
flask-cors==4.0.0
sqlalchemy==2.0.23
paho-mqtt==1.6.1
kafka-python==2.0.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
boto3==1.29.7
azure-storage-blob==12.19.0
google-cloud-storage==2.14.0
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
flask-socketio==5.3.6
eventlet==0.33.3


